+++ api/src/main/java/org/pentaho/dictionary/DictionaryHelper.java
@@ -185,8 +185,10 @@ public class DictionaryHelper {
     typeCategoryMap.put( NODE_TYPE_ENTITY, CATEGORY_ABSTRACT );
     typeCategoryMap.put( NODE_TYPE_FILE, CATEGORY_FIELD_COLLECTION );
     typeCategoryMap.put( NODE_TYPE_JSON_FILE, CATEGORY_FIELD_COLLECTION );
+    typeCategoryMap.put( NODE_TYPE_HDFS_FILE, CATEGORY_FIELD_COLLECTION );
     typeCategoryMap.put( NODE_TYPE_FILE_FIELD, CATEGORY_FIELD );
     typeCategoryMap.put( NODE_TYPE_JSON_FIELD, CATEGORY_FIELD );
+    typeCategoryMap.put( NODE_TYPE_HDFS_FIELD, CATEGORY_FIELD );
     typeCategoryMap.put( NODE_TYPE_LOCATOR, CATEGORY_REPOSITORY );
     typeCategoryMap.put( NODE_TYPE_ROOT_ENTITY, CATEGORY_ABSTRACT );
     typeCategoryMap.put( NODE_TYPE_WEBSERVICE, CATEGORY_DATASOURCE );
@@ -199,6 +201,7 @@ public class DictionaryHelper {
     registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_DATASOURCE, NODE_TYPE_EXTERNAL_CONNECTION );
     registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_DATA_TABLE, null );
     registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_DATA_COLUMN, null );
+    registerEntityType( LINK_CONTAINS_CONCEPT, NODE_TYPE_DATA_COLUMN, NODE_TYPE_DATA_TABLE );
     registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_MONGODB_CONNECTION, NODE_TYPE_EXTERNAL_CONNECTION );
     registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_MONGODB_COLLECTION, null );
     registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_JOB, null );
@@ -213,6 +216,9 @@ public class DictionaryHelper {
     registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_JSON_FILE, NODE_TYPE_FILE );
     registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_JSON_FIELD, NODE_TYPE_FILE_FIELD );
     registerEntityType( LINK_CONTAINS_CONCEPT, NODE_TYPE_JSON_FIELD, NODE_TYPE_JSON_FILE );
+    registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_HDFS_FILE, NODE_TYPE_FILE );
+    registerEntityType( LINK_PARENT_CONCEPT, NODE_TYPE_HDFS_FIELD, NODE_TYPE_FILE_FIELD );
+    registerEntityType( LINK_CONTAINS_CONCEPT, NODE_TYPE_HDFS_FIELD, NODE_TYPE_HDFS_FILE );
   }
 
   /**
