+++ api/src/main/resources/org/pentaho/metaverse/api/messages/messages.properties
@@ -16,7 +16,7 @@ ERROR.ParentJobMeta.IsNull=Parent JobMeta is null!
 #ERROR.NoContentForFile=Could not get file contents for {0}
 #ERROR.MetaverseReader.IsNull=MetaverseReader is null!
 #ERROR.AnalyzingDocument=Error while analyzing document: {0}.
-#ERROR.BackingGraph.MustImplement.KeyIndexableGraph=The configured backing graph must implement com.tinkerpop.blueprints.KeyIndexableGraph
+ERROR.BackingGraph.MustImplement.KeyIndexableGraph=The configured backing graph must implement com.tinkerpop.blueprints.KeyIndexableGraph
 #ERROR.BaseLocator.ScanAlreadyExecuting=The system is already scanning. Only one scan can execute at a time.
 #ERROR.RepositoryLocator.ScanAbortedNoRepo=Scan aborted. Error locating repository.
 #ERROR.ProcessFileFailed=File could not be processed: {0}
